+++ lib/controller/drugsController.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:master_package/models/drug_by_generic_model.dart';
+import 'package:master_package/models/medicine/drug_by_generic_model.dart';
 
 class DrugController extends GetxController {
   var _drugs = {}.obs;
