--- lib/models/drug_by_generic_model.dart
@@ -1,173 +1,0 @@
-class DrugsByGeneric {
-  final int id;
-  final String drugName;
-  final String drugDescription;
-  final String dImage;
-  final String dPrice;
-
-  DrugsByGeneric(
-      {required this.id,
-      required this.drugName,
-      required this.drugDescription,
-      required this.dImage,
-      required this.dPrice});
-
-  static List<DrugsByGeneric> drugs = [
-    DrugsByGeneric(
-        id: 1,
-        drugName: "Abaloparatide Injection(Tymlos)",
-        drugDescription: "Description 1",
-        dImage: "drug Image",
-        dPrice: " \$100"),
-    DrugsByGeneric(
-        id: 2,
-        drugName: "Baclofen(Kemostro)",
-        drugDescription: "Description 2",
-        dImage: "drug Image",
-        dPrice: " \$110"),
-    DrugsByGeneric(
-        id: 3,
-        drugName: "Cabergoline(Dostinex)",
-        drugDescription: "Description 3",
-        dImage: "drug Image",
-        dPrice: " \$120"),
-    DrugsByGeneric(
-        id: 4,
-        drugName: "Daclizumab(Zenapax)",
-        drugDescription: "Description 4",
-        dImage: "drug Image",
-        dPrice: " \$130"),
-    DrugsByGeneric(
-        id: 5,
-        drugName: "Econazole Nitrate(Spectazol)",
-        drugDescription: "Description 5",
-        dImage: "drug Image",
-        dPrice: " \$140"),
-    DrugsByGeneric(
-        id: 6,
-        drugName: "Falmina(Levonorgestral and Ethinyl Estradiol Tablets)",
-        drugDescription: "Description 6",
-        dImage: "drug Image",
-        dPrice: " \$150"),
-    DrugsByGeneric(
-        id: 7,
-        drugName: "Gabapentin Tablets(Gralise)",
-        drugDescription: "Description 7",
-        dImage: "drug Image",
-        dPrice: " \$160"),
-    DrugsByGeneric(
-        id: 8,
-        drugName: "Haemophilus b conjugate Vaccine(ActHIB)",
-        drugDescription: "Description 8",
-        dImage: "drug Image",
-        dPrice: " \$170"),
-    DrugsByGeneric(
-        id: 9,
-        drugName: "Ibsrela(Tenapenor Tablets)",
-        drugDescription: "Description 9",
-        dImage: "drug Image",
-        dPrice: " \$180"),
-    DrugsByGeneric(
-        id: 10,
-        drugName: "Jadenu(Deferasirox Tablets)",
-        drugDescription: "Description 10",
-        dImage: "drug Image",
-        dPrice: " \$190"),
-    DrugsByGeneric(
-        id: 11,
-        drugName: "K-LOR(Potassium Chloride)",
-        drugDescription: "Description 11",
-        dImage: "drug Image",
-        dPrice: " \$200"),
-    DrugsByGeneric(
-        id: 12,
-        drugName: "Lenvatinib",
-        drugDescription: "Description 12",
-        dImage: "drug Image",
-        dPrice: " \$210"),
-    DrugsByGeneric(
-        id: 13,
-        drugName: "Medroxyprogesterone",
-        drugDescription: "Description 13",
-        dImage: "drug Image",
-        dPrice: " \$220"),
-    DrugsByGeneric(
-        id: 14,
-        drugName: "Naprosyn",
-        drugDescription: "Description 14",
-        dImage: "drug Image",
-        dPrice: " \$230"),
-    DrugsByGeneric(
-        id: 15,
-        drugName: "Omeprazole",
-        drugDescription: "Description 15",
-        dImage: "drug Image",
-        dPrice: " \$240"),
-    DrugsByGeneric(
-        id: 16,
-        drugName: "Paracetamol",
-        drugDescription: "Description 16",
-        dImage: "drug Image",
-        dPrice: " \$250"),
-    DrugsByGeneric(
-        id: 17,
-        drugName: "Quazepam",
-        drugDescription: "Description 17",
-        dImage: "drug Image",
-        dPrice: " \$260"),
-    DrugsByGeneric(
-        id: 18,
-        drugName: "Ranitidine",
-        drugDescription: "Description 18",
-        dImage: "drug Image",
-        dPrice: " \$270"),
-    DrugsByGeneric(
-        id: 19,
-        drugName: "Selenium",
-        drugDescription: "Description 19",
-        dImage: "drug Image",
-        dPrice: " \$280"),
-    DrugsByGeneric(
-        id: 20,
-        drugName: "Temovate",
-        drugDescription: "Description 20",
-        dImage: "drug Image",
-        dPrice: " \$290"),
-    DrugsByGeneric(
-        id: 21,
-        drugName: "Ultravist",
-        drugDescription: "Description 21",
-        dImage: "drug Image",
-        dPrice: " \$300"),
-    DrugsByGeneric(
-        id: 22,
-        drugName: "Vascepa",
-        drugDescription: "Description 22",
-        dImage: "drug Image",
-        dPrice: " \$310"),
-    DrugsByGeneric(
-        id: 23,
-        drugName: "Welireg",
-        drugDescription: "Description 23",
-        dImage: "drug Image",
-        dPrice: " \$320"),
-    DrugsByGeneric(
-        id: 24,
-        drugName: "Xeomin",
-        drugDescription: "Description 24",
-        dImage: "drug Image",
-        dPrice: " \$330"),
-    DrugsByGeneric(
-        id: 25,
-        drugName: "Yarrow",
-        drugDescription: "Description 25",
-        dImage: "drug Image",
-        dPrice: " \$340"),
-    DrugsByGeneric(
-        id: 26,
-        drugName: "Zarxio",
-        drugDescription: "Description 26",
-        dImage: "drug Image",
-        dPrice: " \$350"),
-  ];
-}
