+++ lib/screens/doctors/doctors_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:master_package/controller/docController.dart';
-import 'package:master_package/models/docs_model.dart';
+import 'package:master_package/models/doctor/docs_model.dart';
 
 class DoctorsListScreen extends StatelessWidget {
   const DoctorsListScreen({Key? key}) : super(key: key);
